// mobile/src/module_bindings/types.ts
export const User = __t.object("User", {
  identity: __t.identity(),
  username: __t.string(),
  displayName: __t.string(),
  avatarBytes: __t.option(__t.byteArray()),
  get status() {
    return UserStatus;
  },
  createdAt: __t.timestamp(),
  bio: __t.option(__t.string()),
  profileColor: __t.option(__t.string()),
});
export type User = __Infer<typeof User>;

export const UserPresence = __t.object("UserPresence", {
  identity: __t.identity(),
  get status() {
    return UserStatus;
  },
  changedAt: __t.timestamp(),
});
export type UserPresence = __Infer<typeof UserPresence>;

// The tagged union or sum type for the algebraic type `UserStatus`.
export const UserStatus = __t.enum("UserStatus", {
  Online: __t.unit(),
  Away: __t.unit(),
  DoNotDisturb: __t.unit(),
  Offline: __t.unit(),
});
export type UserStatus = __Infer<typeof UserStatus>;